feat(gui): add date shortcuts and UI icons to ResultsController
- Implemented quick-action buttons for date ranges (Today, Yesterday, Week, Month).
- Added icons for all result labels and financial fields.
- Implemented clear button action.
- Added a listener to automatically set final date based on start date.

+++ schema.sql
@@ -9,7 +9,7 @@ CREATE TABLE IF NOT EXISTS orders (
     ifood_payment_value DECIMAL(10, 2),
     ifood_comission DECIMAL(10, 2),
     ifood_direct_payment_value DECIMAL(10, 2),
-    service_fee INT,
+    service_fee DECIMAL(10, 2),
     category VARCHAR(50)
 );
